It compiles now!
Nothing works yet, but hey, progress! Next up is getting the initial
connection and user registration working, along with starting to keep
track of network state. Unit tests will be used for this, along with
maybe spinning up ircds as part of testing to connect to rather than
relying on public services that we could be banned or restricted from.