WIP split server stuff to its own assembly
FakeServer was getting sufficiently complex that keeping it all in one
file seemed like not a great move. The initial scope for Netwolf.Server
is still only to implement enough of an ircd for use in testing, and the
actual daemon bits will be implemented in different projects using
Netwolf.Server as a base (shared) assembly.