Add client-side SASL support
Supporting EXTERNAL, SCRAM-*, and PLAIN for now. I looked into the
various ECC mechs (NIST256P and X25519), however platform support is
flaky right now for X25519 and I'd rather see how future versions of
.NET improve the APIs around these curves. In any event, it seems that
neither will be supported on browser (WASM) unless I go with a 3rd-party
crypto library. The only native C# one is BouncyCastle, though, and it
is a massive dependency for the little I'd be using it for. libsodium is
way faster but has no official C#/.NET bindings plus I'd need to figure
out how to ship it as a dependency for the various platforms.

In all, would rather wait to see X25519 support show up in the standard
library before embarking on this.