More work on bot permissions management
Introduce two new keyed service types (keyed on bot name):
IAccountProvider and IPermissionProvider. Rather than relying on a chain
of IContextAugmenters, this allows for registration of providers via a
new callback on AddBot (via extension methods provided to IBotBuilder).

If multiple account providers are given, the first one that returns a
positive result is used. If multiple permission providers are given, all
permissions from all providers will be merged together to determine the
sender's full permission set.

As of now, there are no implementations of either of these two new
services, nor any extension methods on IBotBuilder to register
providers. A future commit will address both of these with a couple of
generally-useful providers.

Also change from an options snapshot to an options monitor on Bot so
that we have hot-reloadable configuration.

I'm still not happy with the constructor injection situation for Bot and
will likely fiddle around with that to see if I can arrive at something
that is less likely to change all of the time as I add new service
dependencies to the base Bot class. However this might be as good as I
can get using the built-in DI container.