Split waiting for results into a separate method
This introduces two custom awaitables: DeferredCommand and
ConfiguredDeferredCommand. There are guards in place to ensure that
awaiting one of these will only send the command exactly once even if it
is awaited multiple times; for that reason a DeferredCommand can only be
configured zero or one times, and cannot be awaited directly if
configured.

SendAsync overloads were removed and some Bot methods marked internal
instead of private to allow this to work. Now fluent chaining happens
via WithReply or WithReplies depending on if the caller is interested in
just a single value or multiple replies.