Pass the token provided to deferred command in more places
Now when using a configured deferred command as an IAsyncEnumerable, the
CancellationToken passed to the constructor will be combined with any
token passed directly to the enumerator so that e.g. timeouts aren't
lost in such instances.

If ToObservable() is used instead then it's the caller's responsibility
to pass any necessary tokens along as part of the Rx pipeline.