Move state tracking to Transport
Originally I thought I would need to augment state tracking with
framework-specific info, and that may still be the case, however for
various reasons it really needed to live in Transport so we could at
least keep track of ourselves.

State is now fully immutable, with the collections and objects being
replaced on change. This in turn makes it fully thread-safe without
needing to worry about locks or performance impacts when doing read
operations in user code.