WIP loadable plugins
Doesn't fully work yet and I still need to think about how wide the
interop layer between plugin assemblies and the base assembly will be.
Ideally all interaction goes through IPlugin (to pull data from the
plugin) or IPluginHost (for the plugin to pull data from the app or to
push data to the app).

Right now only Netwolf assemblies are shared, all other assemblies are
unique to the plugin. Because of this, I also need to think on how to
ensure that strong refs to plugin assemblies don't leak into the main
app (and therefore prevent plugin unloading/reloading).