More work on plugins
Command hooks are almost functional now. Still need to figure out how to
expose some sort of limited INetwork for plugins to query and use (via
PluginHost as a wrapper). This may require a dependency inversion for
PluginFramework to depend on Transport instead of the other way around,
or perhaps something else.