Make Transport a dependency of PluginFramework
Instead of the other way around. This involves moving a whole bunch of
types back to Transport so that library can be self-contained. This also
unlocks a lot of power in PluginFramework in doing things with networks
and operating on network state that wasn't feasible with the dependency
ordering reversed.

This does meant that PluginFramework is no longer useful standalone
outside of Netwolf, but that wasn't a primary goal for it anyway.