Refactor a bunch of stuff
Moving a decent chunk of logic around DeferredCommand from BotFramework
to the base Transport library, so that PluginFramework can make use of
it.

Internal Network should probably make use of this new feature where
sensible to co-locate some logic, and Network also needs to probably be
moved to a different namespace and split up into multiple files (it's
already a partial class and it's getting huge).