Shift some commands out of Network
More to come, but I want to ensure there's a test suite that passes
before I move things out to validate it was done correctly. The test
suite for USERHOST isn't comprehensive yet and I might just remove the
hostname update portion of that and add in a native method to join
channels that does a WHO(X) and uses that for hostname updates. As a
bonus, that would allow me to track library-issued WHOX at this layer
rather than needing to implement it separately in the bot layer, client
layer, etc.