Lots of changes to Network
This migrates a bunch more commands out into observers on the
CommandReceived stream. In doing so, I discovered some deadlocks during
unit tests that led me to realize that the default IObservable scheduler
was scheduling everything on the current thread. This gave nice
properties such as ensuring that each received command is processed
in-order, but deadlocks aren't very nice and I wasn't able to avoid them
with purely changes in the Network class (changes outside are infeasible
since I can't control what consumers of the library do with their
observers).

Because of this, I moved the default scheduler to its own dedicated
thread. This prevented the deadlock, but now I had a problem that
exceptions thrown during observer callbacks were now going into that
scheduler thread and then bubbling up to the application-level unhandled
exception handler (i.e. program crash). These exceptions now disconnect
the network and pass the exception along to the NetworkDisconnected
event, but do not crash the process.

Changing the scheduler to its own thread also exposed that
UnsafeReceiveRaw is non-blocking, as was already documented. However,
unit tests were assuming the default in-thread scheduler such that it
turned into a blocking method instead. ReceiveLineForUnitTests was
refactored to convert the non-blocking method into a blocking one, that
can also listen in for exceptions as appropriate.

Command listeners for built-in command handling got split into sync and
async listeners, so that we don't invoke heavy async machinery for the
majority of callbacks which operate synchronously anyway. It also let me
clean those classes up by removing the various ugly `return
Task.CompletedTask` statements.

Not done as part of these changes, but noticed as part of them, was that
the NetworkEvents class seemed to be in a really weird place. I think I
was using the global behavior of it at one point in time, but I'm no
longer doing that and the events really would fit better directly on
INetwork (because they're related to the connection/disconnection state
of the network, part of the single reponsibility I'm trying to move the
class over to). A future change will merge this back into the class.